src/info: structs: switch to a lower type for response_code
It makes no sense to use i64 as a response code as the response codes are not of huge integers.

Suggested-by: PeroSar <perosar1111@gmail.com>

+++ src/info/structs/mod.rs
@@ -39,6 +39,6 @@ pub struct SpamPredict {
 #[derive(Deserialize, Debug)]
 pub struct ApiResp {
     pub success: bool,
-    pub response_code: i64,
+    pub response_code: i16,
     pub results: Response
 }
